+++ src/village-manager.cpp
@@ -6,7 +6,7 @@ VillageManager::VillageManager()
 
 }
 
-int VillageManager::addVillage(Faction faction)
+long VillageManager::addVillage(Faction faction)
 {
 	villageList.push_back(new Village(faction));
 	return villageList.size();
@@ -14,8 +14,8 @@ int VillageManager::addVillage(Faction faction)
 
 void VillageManager::update()
 {
-	int villageCount = villageList.size();
-	for(int villageIndex = 0; villageIndex < villageCount; villageIndex ++)
+	long villageCount = villageList.size();
+	for(long villageIndex = 0; villageIndex < villageCount; villageIndex ++)
 	{
 		VillageReference village = villageList[villageIndex];
